Add a pre-order traverser and a benchmark for it.
I'm fairly certain this or something similar will be very useful for
implementing range queries. This is a starting point.

The first implementation I tried was non-recursive, a misguided
attempt to avoid exceeding the stack size for large traversals. It
turns out goroutines do not use the system stack for function calls,
it can grow as large as it needs to. And, this much simpler recursive
implementation also turns out to be vastly more efficient in time and
space than a non-recursive solution.