Remove OrderedBinaryMap.Cursor and Pos.
There aren't enough use cases for mutation during iteration to warrant
keeping the capability, especially considering how difficult it might
be to implement in practice. Users who need this should be able to
accumulate pending mutations and perform them following the iteration.

This also greatly reduces the size of the API.