Test the test impls more thoroughly.
Trying to find a performance problem with the Delete benchmark, so
making sure the test cases are what they're supposed to be. The
individual benchmarked Delete ops are fast, but the total test run is
very much not. I would have though it was the Clone(), but Put isn't
suffering from the same problem.

However, it still might be Clone() since it looks like the Delete
operations are much, much faster. The number of Clone() invocations
might matter.