Add node-based traversers and use them in pointer-trie Range().
These are in addition to the existing path-based traversers and
associated elements, which have been renamed to have the word "path"
in them. Added unit tests and benchmarks for these as well.

The existing ones are roughly:
  path:      []node                 // path[0] is the root
  adj func:  path -> iter.Seq[node] // nodes adjacent to the path's end
  traverser: root, adj func -> iter.Seq[path]

The new ones are:
  adj func:  node -> iter.Seq[node]
  traverser: root, adj func -> iter.Seq[node]

The motivation is that this particular Range implementation doesn't
need paths to nodes, it only needs the keys built from those paths,
and doing it this way is more efficient. Since you need the actual
node as well to compute adjacency, the internal adjacency functions
and traversers use a *struct{node, key} as the traversal "node".